==> scripts/1 - Load Data.ipynb <==
# load all data and select purchase files

import json
import pandas as pd
import glob
import time
from tqdm import tqdm
import sys,os
import collections
import re
import csv

# Choose store
# project = 'he'
# project = 'pp'
project = 'b'

if project == 'he':
    root_dir = '/Data/herbatica_365d/'
    session_file = 'sessions_all_he.csv'
    purchase = 'dakujeme' 
    
elif project == 'pp':
    root_dir = '/Data/particlepeptides_365d/'
    session_file = 'sessions_all_pp_500.csv'
    purchase = 'potvrdenie-objednavky|order-confirmation'
    
elif project == 'b':
    root_dir = '/Data/barefootky_365d/'
    session_file = 'sessions_all_b_500.csv'
    purchase = '&objednane'

data_file = 'data.json'

# list of all files (sessions)
listOfFiles = list()
for (dirpath, dirnames, filenames) in os.walk(root_dir):
    for file in filenames:        
        if file == data_file:
            listOfFiles += [dirpath + "/" + file]

# save list of all files
if project == 'he':
    f = open(session_file, "w")
    f.writelines(["%s\n" % item  for item in listOfFiles])
    f.close()

if project == 'pp':
    f = open(session_file, "w")
    f.writelines(["%s\n" % item  for item in listOfFiles])
    f.close()
    
if project == 'b':
    f = open(session_file, "w")
    f.writelines(["%s\n" % item  for item in listOfFiles])
    f.close()

# load list of all files 
listOfFiles = list()
with open(session_file, newline = '') as csvfile:
    reader = csv.reader(csvfile)
    for row in reader:
        listOfFiles.append(row[0])

len(listOfFiles) 

# sessions with purchase - dakujeme (he), potvrdenie-objednavky (pp), &objednavka (ba)
buyFiles = list()
for filename in listOfFiles:
    with open(filename, 'r', encoding="utf8") as file:
        file = file.read()
        json_data = json.loads(file)
        pageviews = json_data['pageviews']
        for pv in pageviews:
            try:
                if re.findall(purchase, pv['url']):
                    buyFiles.append(filename)
            except Exception:
                pass

buyFiles = set(buyFiles)

len(buyFiles)

if project == 'he':    
    f = open("sessions_purchase_all_he.csv", "w")
    f.writelines(["%s\n" % item  for item in buyFiles])
    f.close()
    
if project == 'pp':
    f = open("sessions_purchase_all_pp_500.csv", "w")
    f.writelines(["%s\n" % item  for item in buyFiles])
    f.close()
    
if project == 'b':
    f = open("sessions_purchase_all_b_500.csv", "w")
    f.writelines(["%s\n" % item  for item in buyFiles])
    f.close()
