==> scripts/2 - Parsing.ipynb <==
import json
import pandas as pd
import glob
import time
from tqdm import tqdm
import sys,os
import collections
import re
import csv
from datetime import datetime

# project = 'he'
# project = 'pp'
project = 'b'

if project == 'he':
    root_dir = '/Data/herbatica_365d/'
    session_file = 'sessions_all_he.csv'
    buyers_file = 'sessions_purchase_all_he.csv'
    session_buyers = "sessions_buyers_all_he.csv"
    folder = 'csv_he'

elif project == 'pp':
    root_dir = '/Data/particlepeptides_365d/'
    session_file = 'sessions_all_pp_500.csv'
    buyers_file = 'sessions_purchase_all_pp_500.csv'
    session_buyers = "sessions_buyers_all_pp_500.csv"
    folder = 'csv_pp_500'

elif project == 'b':
    root_dir = '/Data/barefootky_365d/'
    session_file = 'sessions_all_b_500.csv'
    buyers_file = 'sessions_purchase_all_b_500.csv'
    session_buyers = "sessions_buyers_all_b_500.csv"
    folder = 'csv_b_500'

data_file = 'data.json'

# load all files
listOfFiles = list()
with open(session_file, newline = '') as csvfile:
    reader = csv.reader(csvfile)
    for row in reader:
        listOfFiles.append(row[0])

# load purchase sessions
listOfBuyers = list()
with open(buyers_file, newline = '') as csvfile:
    reader = csv.reader(csvfile)
    for row in reader:
        listOfBuyers.append(row[0])

session_excludes = [ 
    'ip',
    'id',
    'isEu', 
    'latitude', 
    'longitude', 
    'projectId', 
    'useragent', 
    'version',
    'engine',
    'pageviews',
]
session_normalized_excludes = [
    'browser.family.name', 
    'browser.family.version', 
    'browser.type', 
    'browser.version',
    'browser.using.name',
    'browser.using.version',
    'device.manufacturer', 
    'device.model',
    'device.subtype',
    'device.series',
    'os.alias',
    'os.version',
    'os.version.value',
    'os.version.alias',
    'os.version.nickname',
    'os.edition',
    'os.family',
    'user.projectId',
    'user.number',
    'user.recorded'
]
pageview_excludes = [
    'screenWidth',
    'screenHeight',
    'width',
    'height',
]

df_sessions = pd.DataFrame()
df_users = pd.DataFrame()
df_pageviews = pd.DataFrame()

df_load = pd.DataFrame()
df_click = pd.DataFrame()
df_input = pd.DataFrame()
df_mouse_move = pd.DataFrame()
df_mouse_click = pd.DataFrame()
df_scroll_move = pd.DataFrame()
df_scrandom = pd.DataFrame()
df_wild_mouse = pd.DataFrame()
df_rage_click = pd.DataFrame()

action_types = ["input","load","click","mouse-click","rage-click","scroll-move","mouse-move","wild-mouse","scrandom"]

df_includes = {
    'input_includes':["pageview_id","session_id","id","type","at","args.duration","args.type","args.name"],
    'load_includes':["pageview_id","session_id","id","type","at",'args.firstInsertAt','args.domContentLoaded','args.firstPaint',
                   'args.firstContentfulPaint','args.loadEvent','args.firstInputDelay','args.timeToInteractive'],
    'click_includes':["pageview_id","session_id","id","type","at","args.double","args.type","args.text","args.position.x","args.position.y","args.target.left",
                    "args.target.top","args.target.width",'args.target.height'],
    'mouse-click_includes':["pageview_id","session_id","id","type","at","args.button","args.down.x","args.down.y","args.down.at","args.up.x","args.up.y",
                          "args.up.at","args.pause","args.duration","args.distance"],
    'rage-click_includes':["pageview_id","session_id","id","type","at","args.duration","args.center.x","args.center.y","args.intensity"],
    'scroll-move_includes':["pageview_id","session_id","id","type","at","args.duration","args.distance","args.id","args.source.pageLeft","args.source.pageTop","args.source.at",
                          "args.destination.pageLeft","args.destination.pageTop","args.destination.at","args.direction","args.distance2",
                          "args.vx","args.vy","args.v","args.a","args.jerk"],
    'mouse-move_includes':["pageview_id","session_id","id","type","at","args.duration","args.distance","args.source.at","args.destination.at","args.direction","args.distance2",
                         "args.vx","args.vy","args.v","args.a","args.jerk","args.source.x","args.source.y","args.destination.x",
                         "args.destination.y","args.straightness","args.jitter"],
    'wild-mouse_includes':["pageview_id","session_id","id","type","at","args.duration"],
    'scrandom_includes':["pageview_id","session_id","id","type","at","args.duration"]   
}

# list of buyers - user id
listOfUserBuyers = list()
for filename in listOfBuyers:
    with open(filename, 'r', encoding="utf8") as file:
        file = file.read()
        json_data = json.loads(file)           
        listOfUserBuyers.append(json_data['user']['id'])

listOfUserBuyers = (set(listOfUserBuyers))

len(listOfUserBuyers)

# all sessions of buyers
listOfSessions = list()
for filename in listOfFiles:
    with open(filename, 'r', encoding="utf8") as file:
        file = file.read()
        json_data = json.loads(file)
        if json_data['user']['id'] in listOfUserBuyers:
            listOfSessions.append(filename)

f = open(session_buyers, "w")
f.writelines(["%s\n" % item  for item in listOfSessions])
f.close()

# load all sessions of buyers
listOfSessions = list()
with open(session_buyers, newline = '') as csvfile:
    reader = csv.reader(csvfile)
    for row in reader:
        listOfSessions.append(row[0])

len(listOfSessions) 

# pageview parsing
for filename in listOfSessions:
    with open(filename, 'r', encoding="utf8") as file:
        file = file.read()
        json_data = json.loads(file)        
        parsed_pageviews = pd.DataFrame(json_data['pageviews'])
        parsed_pageviews = parsed_pageviews.drop(columns=pageview_excludes, axis=1, errors='ignore')
        parsed_pageviews['session_id'] = json_data['sessionId'] # add session_id to pageviews to connect pageview to a specific session  
        try:
            parsed_pageviews = parsed_pageviews[parsed_pageviews['baked'].str.len() > 0] # not empty baked only 
            if not parsed_pageviews.empty:
                # sessions
                session_data = {k: json_data[k] for k in set(list(json_data.keys())) - set(session_excludes)}
                session_data = pd.json_normalize(session_data) # split into columns
                session_data = session_data.drop(columns=session_normalized_excludes, axis=1, errors='ignore')                     
                df_sessions = df_sessions.append(session_data, ignore_index=True, sort=False) # session append

                # save baked from pageview into parsed_events
                parsed_events = pd.DataFrame(parsed_pageviews[['id', 'session_id', 'baked']])
                parsed_pageviews = parsed_pageviews.drop(columns=['baked']) 

                # parsing pageviews                
                df_pageviews = df_pageviews.append(parsed_pageviews, ignore_index=True, sort=False) # pageview append
                parsed_events.rename(columns={'id':'pageview_id'}, inplace=True)
                parsed_events = parsed_events.explode('baked') # explode baked objects into rows

                # explode baked
                for key in parsed_events.iloc[0]['baked'].keys():
                    parsed_events[key] = parsed_events.baked.apply(lambda b: b[key] if type(b) is dict else None)

                parsed_events = parsed_events.drop(columns=['baked', 'modifiers'])
                
                for type_name in action_types:
                    type_includes = df_includes[f'{type_name}_includes']

                    type_data = parsed_events[parsed_events.type == type_name]
                    type_data = type_data.join(pd.json_normalize(type_data.args.values).add_prefix('args.'))
                    type_data = type_data.drop(columns=['args','type'], axis=1, errors='ignore')
                    type_data = type_data[type_data.columns & type_includes]

                    if (type_name == 'load'):
                        df_load = df_load.append(type_data, ignore_index=True)
                    elif (type_name == 'input'):
                        df_input = df_input.append(type_data, ignore_index=True)
                    elif (type_name == 'click'):
                        df_click = df_click.append(type_data, ignore_index=True)
                    elif (type_name == 'mouse-click'):
                        df_mouse_click = df_mouse_click.append(type_data, ignore_index=True)
                    elif (type_name == 'rage-click'):
                        df_rage_click = df_rage_click.append(type_data, ignore_index=True)
                    elif (type_name == 'scroll-move'):
                        df_scroll_move = df_scroll_move.append(type_data, ignore_index=True) 
                    elif (type_name == 'mouse-move'):
                        df_mouse_move = df_mouse_move.append(type_data, ignore_index=True)
                    elif (type_name == 'wild-mouse'):
                        df_wild_mouse = df_wild_mouse.append(type_data, ignore_index=True)
                    elif (type_name == 'scrandom'):
                        df_scrandom = df_scrandom.append(type_data, ignore_index=True)                   

        except KeyError:
            print("Baked empty")

def convertMS(ms):
    return datetime.fromtimestamp(ms/1000.0)

# SPRACOVANIE - result: 
# referrer, started_at, day, session_id, country, city, region, continent, user_id, device.type, browser.name, os.name, started_at_time, 
# ses_id

df_sessions = df_sessions.rename(columns={'startedAt': 'started_at', 'sessionId': 'session_id', 'user.id': 'user_id'})
df_sessions['started_at_time'] = df_sessions['started_at'].apply(lambda ms: convertMS(ms))
df_sessions['ses_id'] = pd.factorize(df_sessions['session_id'])[0] + 1 # replace sessionID with numbers (1...N)

# USERS

df_users = pd.DataFrame(df_sessions[['user_id', 'started_at', 'started_at_time']])
df_users['u_id'] = pd.factorize(df_users['user_id'])[0] + 1 

# session count for each user
sessions_count = pd.DataFrame(df_users['u_id'].value_counts()) # user_id count
sessions_count.rename(columns={'u_id': 'n_sessions'}, inplace=True) 
df_users.drop_duplicates('u_id', keep='first', inplace=True)

sessions_count.head(3)

df_users = pd.merge(df_users, sessions_count, left_on='u_id', right_index=True, how='left')

df_users.sort_values('n_sessions', ascending = False).head(6)

# PAGEVIEWS - result: 
# pageview_id, url, referrer, started_at, duration, inactivity, session_id, started_at_time, pv_id, ses_id, user_id

df_pageviews = df_pageviews.rename(columns={'id': 'pageview_id', 'startedAt': 'started_at'})
df_pageviews['started_at_time'] = df_pageviews['started_at'].apply(lambda ms: convertMS(ms))
df_pageviews['pv_id'] = pd.factorize(df_pageviews['pageview_id'])[0] + 1
df_pageviews.head(4)

df_pageviews = pd.merge(
    df_pageviews,
    df_sessions[['session_id', 'ses_id', 'user_id']],
    on='session_id',
    how='left'
)

# SPRACOVANIE INACTIVITIES

df_inactivities = pd.DataFrame(df_pageviews[['pageview_id','pv_id','session_id','ses_id','user_id','started_at', 'started_at_time','inactivity']])
df_pageviews = df_pageviews.drop(columns=['inactivity'], axis=1, errors='ignore')
df_inactivities = df_inactivities.explode('inactivity')

# inactivity explode
for key in df_inactivities.iloc[0]['inactivity'].keys():
    df_inactivities[key] = df_inactivities.inactivity.apply(lambda b: b[key] if type(b) is dict else None)

df_inactivities = df_inactivities.drop(columns=['inactivity'])
df_inactivities.rename(columns={'start': 'inactivity_start'}, inplace=True)
df_inactivities.rename(columns={'end': 'inactivity_end'}, inplace=True)

# sum of inactivities
df_inactivities['inactivity_sum'] = df_inactivities['inactivity_end'] - df_inactivities['inactivity_start']

# sum of inactivities for user
df_inactivities = df_inactivities.sort_values(by=['user_id', 'started_at'])
df_inactivities['user_total_inactivity'] = df_inactivities.groupby('user_id')['inactivity_sum'].expanding(1).sum().values
df_inactivities['user_last_total_inactivity'] = df_inactivities.groupby('user_id')['inactivity_sum'].expanding(1).sum().values - df_inactivities['inactivity_sum']

# sum of inactivities for session
df_inactivities = df_inactivities.sort_values(by=['session_id', 'started_at'])
df_inactivities['session_total_inactivity'] = df_inactivities.groupby('session_id')['inactivity_sum'].expanding(1).sum().values
df_inactivities['session_last_total_inactivity'] = df_inactivities.groupby('session_id')['inactivity_sum'].expanding(1).sum().values - df_inactivities['inactivity_sum']

# sum of inactivities for pageview 
df_inactivities = df_inactivities.sort_values(by=['pageview_id', 'started_at']) 
df_inactivities['pageview_total_inactivity'] = df_inactivities.groupby('pageview_id')['inactivity_sum'].expanding(1).sum().values
df_inactivities['pageview_last_total_inactivity'] = df_inactivities.groupby('pageview_id')['inactivity_sum'].expanding(1).sum().values - df_inactivities['inactivity_sum']

# for each pageview compute effective_duration based on df_inactivities['pageview_total_inactivity']
df_inactivities = df_inactivities.sort_values(by=['pageview_id', 'started_at'])

df_pageviews = pd.merge(
    df_pageviews,
    df_inactivities[['pageview_id', 'pageview_total_inactivity']],
    how='left',
    on='pageview_id'
)

df_pageviews.drop_duplicates('pageview_id', keep='last', inplace=True) 
df_pageviews['effective_duration'] = df_pageviews['duration'] - df_pageviews['pageview_total_inactivity']
df_pageviews = df_pageviews.drop(columns=['pageview_total_inactivity'])

# for each session compute effective_duration based on df_pageviews['effective_duration']
df_pageviews = df_pageviews.sort_values(by=['session_id', 'started_at'])
df_sessions = df_sessions.sort_values(by=['session_id', 'started_at'])

df_sessions = pd.merge(
    df_sessions,
    df_pageviews[['session_id', 'effective_duration']],
    how='left',
    on='session_id'
)

df_sessions['pageviews_effective_duration'] = df_sessions.groupby('session_id')['effective_duration'].expanding(1).sum().values
df_sessions.drop_duplicates('session_id', keep='last', inplace=True)
df_sessions = df_sessions.drop(columns=['effective_duration'])
df_sessions = df_sessions.rename(columns={'pageviews_effective_duration': 'effective_duration'})

# for each session compute effective_time_from_start based on effective_duration sessions 
df_sessions = df_sessions.sort_values(by=['user_id', 'started_at'])
df_sessions['effective_time_from_start'] = df_sessions.groupby('user_id')['effective_duration'].expanding(1).sum().values - df_sessions['effective_duration']

# for each pageview compute effective_time_from_start based on effective_duration pageviews
df_pageviews = df_pageviews.sort_values(by=['user_id', 'started_at'])
df_pageviews['effective_time_from_start'] = df_pageviews.groupby('user_id')['effective_duration'].expanding(1).sum().values - df_pageviews['effective_duration']

# SPRACOVANIE BAKED

def add_data_from_pageview(dataframe):
    
    dataframe = pd.merge(
        dataframe,
        df_pageviews[['pageview_id', 'pv_id', 'user_id', 'effective_time_from_start']],
        on='pageview_id',
        how='left'
    )

    dataframe = dataframe.rename(columns={'effective_time_from_start': 'pageview_effective_time_from_start'})
    
    return dataframe

df_input = add_data_from_pageview(df_input)
df_load = add_data_from_pageview(df_load)
df_click = add_data_from_pageview(df_click)
df_mouse_click = add_data_from_pageview(df_mouse_click)
df_rage_click = add_data_from_pageview(df_rage_click)
df_mouse_move = add_data_from_pageview(df_mouse_move)
df_scroll_move = add_data_from_pageview(df_scroll_move)
df_wild_mouse = add_data_from_pageview(df_wild_mouse)
df_scrandom = add_data_from_pageview(df_scrandom)

# for all events compute effective_time_from_start based on pageviews duration and inactivities length
def count_effective_time_from_start(dataframe):

    dataframe = dataframe.reset_index().rename(columns={'index': 'event_id'})

    dataframe = pd.merge(
        dataframe,
        df_inactivities[['pageview_id', 'pv_id','inactivity_end', 'pageview_total_inactivity']],
        on=['pageview_id','pv_id'],
        how='left'
    )

    dataframe = dataframe[dataframe['inactivity_end'] <= dataframe['at']] 
    dataframe.drop_duplicates('event_id', keep='last', inplace=True) 
    dataframe['effective_time_from_start'] = dataframe['pageview_effective_time_from_start'] + dataframe['at'] - dataframe['pageview_total_inactivity']
    dataframe = dataframe.drop(columns=['inactivity_end', 'pageview_total_inactivity', 'pageview_effective_time_from_start'])
    dataframe = dataframe.set_index('event_id')
    
    return dataframe

df_input = count_effective_time_from_start(df_input)
df_load = count_effective_time_from_start(df_load)
df_click = count_effective_time_from_start(df_click)
df_mouse_click = count_effective_time_from_start(df_mouse_click)
df_rage_click = count_effective_time_from_start(df_rage_click)
df_mouse_move = count_effective_time_from_start(df_mouse_move)
df_scroll_move = count_effective_time_from_start(df_scroll_move)
df_wild_mouse = count_effective_time_from_start(df_wild_mouse)
df_scrandom = count_effective_time_from_start(df_scrandom)

df_inactivities.to_csv(folder + '/df_inactivities_all.csv', index=False, encoding='utf-8-sig')

df_input.to_csv(folder + '/df_input_all.csv', index=False, encoding='utf-8-sig')
df_load.to_csv(folder + '/df_load_all.csv', index=False, encoding='utf-8-sig')
df_click.to_csv(folder + '/df_click_all.csv', index=False, encoding='utf-8-sig')
df_mouse_click.to_csv(folder + '/df_mouse_click_all.csv', index=False, encoding='utf-8-sig')
df_rage_click.to_csv(folder + '/df_rage_click_all.csv', index=False, encoding='utf-8-sig')
df_mouse_move.to_csv(folder + '/df_mouse_move_all.csv', index=False, encoding='utf-8-sig')
df_scroll_move.to_csv(folder + '/df_scroll_move_all.csv', index=False, encoding='utf-8-sig')
df_wild_mouse.to_csv(folder + '/df_wild_mouse_all.csv', index=False, encoding='utf-8-sig')
df_scrandom.to_csv(folder + '/df_scrandom_all.csv', index=False, encoding='utf-8-sig')

df_pageviews.to_csv(folder + '/df_pageviews_all.csv', index=False, encoding='utf-8-sig')
df_sessions.to_csv(folder + '/df_sessions_all.csv', index=False, encoding='utf-8-sig')
