==> scripts/3 - Features.ipynb <==
import json
import pandas as pd
import glob
import time
from tqdm import tqdm
import sys,os
import collections
import re
import csv
from datetime import datetime
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.colors as colors
import itertools
import statistics
import scipy.stats as ss
import seaborn as sns
from copy import *

pd.options.display.max_columns = None

# project = 'he'
project = 'pp'
# project = 'ba'

if project == 'he':
    folder = 'csv_he'
    url_regex = r'((?<=https:\/\/www.herbatica.sk\/)\S+)'
    
elif project == 'pp':
    folder = 'csv_pp_500'
    url_regex = r'((?<=https:\/\/particlepeptides.com\/)\S+)'

elif project == 'ba':
    folder = 'csv_b_500'
    url_regex = r'((?<=https:\/\/www.barefootky.sk\/)\S+)'


df_inactivities = pd.read_csv(folder + '/df_inactivities_all.csv')
df_input = pd.read_csv(folder + '/df_input_all.csv')
df_load = pd.read_csv(folder + '/df_load_all.csv')
df_click = pd.read_csv(folder + '/df_click_all.csv')
df_mouse_click = pd.read_csv(folder + '/df_mouse_click_all.csv')
df_rage_click = pd.read_csv(folder + '/df_rage_click_all.csv')
df_mouse_move = pd.read_csv(folder + '/df_mouse_move_all.csv')
df_scroll_move = pd.read_csv(folder + '/df_scroll_move_all.csv')
df_wild_mouse = pd.read_csv(folder + '/df_wild_mouse_all.csv')
df_scrandom = pd.read_csv(folder + '/df_scrandom_all.csv')
df_pageviews = pd.read_csv(folder + '/df_pageviews_all.csv')
df_sessions = pd.read_csv(folder + '/df_sessions_all.csv')

df_pageviews['url1'] = df_pageviews['url'].str.extract(url_regex) # url1 - everything behind https://www.detskapostel.com/
df_pageviews['url2'] = df_pageviews['url1'].str.extract(r'((?<=\/)\S+)') # 2. part from url1 without / at the end
df_pageviews['url3'] = df_pageviews['url2'].str.extract(r'((?<=\/)\S+)') # 2. part from url2 without / at the end

df_pageviews['url1'] = df_pageviews['url1'].str.extract(r'^([^\/]*)') # 1. part from url1 without /
df_pageviews['url2'] = df_pageviews['url2'].str.extract(r'^([^\/]*)') # 1. part from url2 without /
df_pageviews['url3'] = df_pageviews['url3'].str.extract(r'^([^\/]*)') # 1. part from url3 without /    

# do df_pageviews nám pribudli tieto stĺpce:
# - pageview_non_product (True/False) - nonproduct pages
# - pageview_category (True/False) - category pages
# - pageview_product (True/False) - product pages
# - pageview_search - search
# - pageview_buy - purchase completion page
# - pageview_cart - cart page
# - pageview_filter (True/False) - filter in category (pagination and google ads clicks)
#     - if filter = 1, category = 1
#
# pageview_non_product + pageview_category + pageview_product + pageview_search + pageview_buy + pageview_cart = df_pageviews

if (project == 'he'):
    non_product_pages = [
        'vernost',
        'newsletter',
        'blog_herbatica',
        'blog',
        'doprava_platby',
        'certifikaty_produktov',
        'obchodne-podmienky',
        'poou',
        'o-nas',
        'faq',
        'registracia',
        'klient',
        'login',
        'napiste-nam',
        'logout',
        'affiliate-registracia',
        'affiliate-login',
        'affiliate-zabudnute-heslo',
        'user',
    ]
    search = [
        'vyhladavanie',
    ]
    buy = [
        'dakujeme',
    ]
    cart = [
        'kosik',
        'objednavka',
    ]
    pagination = 'strana-'
    filtering = '?'
    
    # nonproduct page / empty url
    df_pageviews.loc[(df_pageviews['url1'].isin(non_product_pages)) | (df_pageviews['url1'].isnull()), 'pageview_non_product'] = True

    df_pageviews['pageview_search'] = df_pageviews['url1'].isin(search) # search page
    df_pageviews['pageview_buy'] = df_pageviews['url2'].isin(buy) # purchase page

    # cart page but not purchase page
    df_pageviews.loc[(df_pageviews['url1'].isin(cart)) & (df_pageviews['pageview_buy'] == False), 'pageview_cart'] = True # cart
    
    # transformation NaN to 0 and float to int
    numeric_values = [
        'pageview_non_product',
        'pageview_search',
        'pageview_buy',
        'pageview_cart'
    ]
    for column in numeric_values:
        df_pageviews[column] = df_pageviews[column].fillna(0)
        df_pageviews[column] = df_pageviews[column].astype(int)

    # check if product page and empty url2 => category page
    df_pageviews.loc[
        (df_pageviews['pageview_non_product'] == False) & 
        (df_pageviews['pageview_cart'] == False) & 
        (df_pageviews['url2'].isnull()) & 
        (df_pageviews['url1'].notna()), 'pageview_category'] = True

    # check if product page and not empty url2 => product page
    df_pageviews.loc[
            (df_pageviews['pageview_non_product'] == False) & 
            (df_pageviews['pageview_buy'] == False) & 
            (df_pageviews['pageview_cart'] == False) & 
            (df_pageviews['url2'].notna()) & 
            (df_pageviews['url1'].notna()), 'pageview_product'] = True

    # if url2 starts 'strana-' or '?', it's not product page, but category page / filter
    df_pageviews.loc[
        (df_pageviews['pageview_product'] == True) & 
        (df_pageviews['pageview_search'] == False) & 
        (df_pageviews['url2'].str.startswith(pagination) | 
         df_pageviews['url2'].str.startswith(filtering)), 'pageview_filter'] = True

    df_pageviews.loc[(df_pageviews['pageview_filter'] == True), 'pageview_category'] = True

    # check if product page and not empty url2 => product page
    df_pageviews.loc[
        (df_pageviews['pageview_product'] == True) & 
        (df_pageviews['url2'].str.startswith(pagination) | 
         df_pageviews['url2'].str.startswith(filtering)), 'pageview_product'] = False

    
if (project == 'pp'):
    category_pages = '16-kupit-peptidy|16-buy-peptides'
    product_detail_pages = [
        'kupit-peptidy',
        'buy-peptides',
    ]
    search_pages = 'search|vyhladavanie'
    buy_pages = 'potvrdenie-objednavky|order-confirmation'
    cart_pages = 'objednavka|order'
    filtering = '?'
    
    df_pageviews['pageview_product'] = df_pageviews['url2'].isin(product_detail_pages)
    df_pageviews['pageview_category'] = df_pageviews['url2'].str.contains(category_pages, regex=True)
    df_pageviews['pageview_search'] = df_pageviews['url2'].str.contains(search_pages, regex=True)
    df_pageviews['pageview_buy'] = df_pageviews['url2'].str.contains(buy_pages, regex=True)
    df_pageviews.loc[
        (df_pageviews['pageview_product'] == False) & 
        (df_pageviews['pageview_category'] == False) & 
        (df_pageviews['pageview_search'] == False) & 
        (df_pageviews['pageview_buy'] == False) & 
        (df_pageviews['url2'].str.contains(cart_pages, regex=True)), 'pageview_cart'] = True
    
    # transformation NaN to 0 and float to int
    numeric_values = [
        'pageview_category',
        'pageview_product',
        'pageview_search',
        'pageview_buy',
        'pageview_cart'
    ]

    for column in numeric_values:
        df_pageviews[column] = df_pageviews[column].fillna(0)
        df_pageviews[column] = df_pageviews[column].astype(int)

    df_pageviews.loc[
        (df_pageviews['pageview_product'] == False) & 
        (df_pageviews['pageview_category'] == False) & 
        (df_pageviews['pageview_buy'] == False) &
        (df_pageviews['pageview_cart'] == False) & 
        (df_pageviews['pageview_search'] == False), 'pageview_non_product'] = True 
    df_pageviews.loc[
        (df_pageviews['pageview_category'] == True) & 
        (df_pageviews['url2'].str.contains(filtering, regex=False)), 'pageview_filter'] = True  

    
if (project == 'ba'):
    category_page = 'kategoria' # url1, category is in url2, product is in url3
    search_page = 'vyhladavanie' # url1
    cart_page = 'nakupny-kosik' # url1
    buy_page = '&objednane' # url2
    filtering = '#' # url3
    
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(category_page)) & 
        (df_pageviews['url3'].isnull()), 'pageview_category'] = True 
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(category_page)) & 
        (df_pageviews['url3'].str.startswith(filtering)), 'pageview_category'] = True 
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(category_page)) & 
        (df_pageviews['url3'].str.startswith(filtering)), 'pageview_filter'] = True 
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(category_page)) & 
        (df_pageviews['url3'].notna()) & 
        (df_pageviews['pageview_filter'].isnull()), 'pageview_product'] = True
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(cart_page)) & 
        (df_pageviews['url2'].str.contains(buy_page, regex=False)), 'pageview_buy'] = True
    df_pageviews.loc[
        (df_pageviews['url1'].str.contains(cart_page)) & 
        (df_pageviews['pageview_buy'].isnull()), 'pageview_cart'] = True
    df_pageviews['pageview_search'] = df_pageviews['url1'].str.contains(search_page)
    
    # transformation NaN to 0 and float to int
    numeric_values = [
        'pageview_category',
        'pageview_product',
        'pageview_search',
        'pageview_buy',
        'pageview_cart'
    ]

    for column in numeric_values:
        df_pageviews[column] = df_pageviews[column].fillna(0)
        df_pageviews[column] = df_pageviews[column].astype(int)

    df_pageviews.loc[
        (df_pageviews['pageview_product'] == False) & 
        (df_pageviews['pageview_category'] == False) & 
        (df_pageviews['pageview_buy'] == False) & 
        (df_pageviews['pageview_cart'] == False) & 
        (df_pageviews['pageview_search'] == False), 'pageview_non_product'] = True 

numeric_values = [
    'pageview_non_product', 'pageview_category', 'pageview_product',
    'pageview_filter', 'pageview_search', 'pageview_buy', 'pageview_cart'
]

for column in numeric_values:
    df_pageviews[column] = df_pageviews[column].fillna(0)
    df_pageviews[column] = df_pageviews[column].astype(int)

# number of pageviews of each type
# df_session['n_buy'] = number of records where df_pageviews['session_id'] == df_sessions['session_id'] and df_pageviews['pageview_buy'] == True

def num_url_per_session(column):
    group = df_pageviews.groupby('session_id')
    
    temp = pd.DataFrame()
    temp = group['pageview_' + column].sum() # sum pageview_buy per session 
    temp = temp.to_frame()
    temp.rename(columns={"pageview_" + column: "n_" + column}, inplace=True) 

    return pd.merge(df_sessions, temp, left_on='session_id', right_index=True, how='left')

df_sessions = num_url_per_session('buy') 
# number of purchase in session should be 1, some are 3 - redirecting from payment gateway

df_sessions.loc[df_sessions['n_buy'] > 0, 'buy'] = True
df_sessions.loc[df_sessions['n_buy'] == 0, 'buy'] = False

df_users = df_sessions.groupby(['user_id','session_id'])[['n_buy']].sum()

df_users.loc[df_users['n_buy'] > 0, 'buy'] = 1
df_users.loc[df_users['n_buy'] == 0, 'buy'] = 0

df_users.sort_values('n_buy').tail(5)

len(df_users.groupby(level='user_id')) # number of unique users

len(df_users[df_users['buy'] > 0].groupby(level='user_id'))

len(df_users[df_users['buy'] == 0].groupby(level='user_id'))

users = df_users.reset_index()

purchase = users[users['buy'] > 0]['user_id'].unique()
no_purchase = users[users['buy'] == 0]['user_id'].unique()

# set(no_purchase.flat) - set(purchase.flat)

# set(no_purchase).intersection(set(purchase)) 

buy_count = users.set_index('user_id').buy.eq(1).sum(level=0).astype(int).reset_index()

buy_count[buy_count['buy'] > 1]

df_sessions = num_url_per_session('product') # n_product - number of products visited per session
df_sessions = num_url_per_session('non_product') # n_non_product
df_sessions = num_url_per_session('category') # n_category 
df_sessions = num_url_per_session('filter') # n_filter
df_sessions = num_url_per_session('search') # n_search
df_sessions = num_url_per_session('cart') # n_cart

def num_per_session(dataframe, table):
    temp = pd.DataFrame(dataframe["session_id"].value_counts()) 
    temp.rename(columns={"session_id": "n_" + table}, inplace=True)
    
    return pd.merge(df_sessions, temp, left_on='session_id', right_index=True, how='left')

df_sessions = num_per_session(df_pageviews, 'pageviews')
df_sessions = num_per_session(df_input, 'input')
df_sessions = num_per_session(df_load, 'load')
df_sessions = num_per_session(df_click, 'click')
df_sessions = num_per_session(df_mouse_click, 'mouse_click')
df_sessions = num_per_session(df_rage_click, 'rage_click')
df_sessions = num_per_session(df_mouse_move, 'mouse_move')
df_sessions = num_per_session(df_scroll_move, 'scroll_move')
df_sessions = num_per_session(df_wild_mouse, 'wild_mouse')
df_sessions = num_per_session(df_scrandom, 'scrandom')

numeric_values = [
    'n_pageviews', 'n_input', 'n_load', 'n_click', 'n_mouse_click', 'n_rage_click',
    'n_mouse_move', 'n_scroll_move', 'n_wild_mouse', 'n_scrandom'
]

for column in numeric_values:
    df_sessions[column] = df_sessions[column].fillna(0)
    df_sessions[column] = df_sessions[column].astype(int)

# number of events per session 
df_sessions['n_events'] = df_sessions['n_input'] + df_sessions['n_load'] + \
            df_sessions['n_click'] + df_sessions['n_mouse_click'] + df_sessions['n_rage_click'] + \
            df_sessions['n_mouse_move'] + df_sessions['n_scroll_move'] + \
            df_sessions['n_wild_mouse'] + df_sessions['n_scrandom']

# top_product (str) - the most visited product per session
# n_top_product (int) - number of visits of the most visited product per session 
# n_unique_product (int) - number of unique products per session

if (project == 'dp' or project == 'he'):
    url_name = 'url2'
elif (project == 'pp' or project == 'ba'):
    url_name = 'url3'

df_pageviews_product = df_pageviews[df_pageviews['pageview_product'] == True][['session_id', url_name]]

# select mode
if (project == 'dp' or project == 'he' or project == 'ba'):
    sessions_top_product = df_pageviews_product.groupby(['session_id'])[url_name].apply(lambda x: pd.Series.mode(x)[0]).to_frame()
elif (project == 'pp'):
    sessions_top_product = df_pageviews_product.groupby(['session_id'])[url_name].agg(pd.Series.mode).to_frame()

sessions_top_product.rename(columns={url_name: 'top_product'}, inplace=True)

df_sessions = pd.merge(
    df_sessions,
    sessions_top_product,
    on='session_id',
    how='left'
)

sessions_n_top_product = df_pageviews_product.groupby(['session_id', url_name])[url_name].count().to_frame()
sessions_n_top_product.rename(columns={url_name: 'n_top_product'}, inplace=True)
sessions_n_top_product = sessions_n_top_product.sort_values(by=['n_top_product'], ascending=False)

df_sessions = pd.merge(
    df_sessions,
    sessions_n_top_product,
    on='session_id',
    how='left'
)

df_sessions.drop_duplicates('session_id', keep='first', inplace=True)
df_sessions = df_sessions.reset_index(drop=True)

sessions_n_unique_product = df_pageviews_product.groupby(['session_id'])[url_name].nunique().to_frame()
sessions_n_unique_product.rename(columns={url_name: 'n_unique_product'}, inplace=True)

df_sessions = pd.merge(
    df_sessions,
    sessions_n_unique_product,
    on='session_id',
    how='left'
)

# top_category (str) - the most visited category per session
# n_top_category (int) - number of visits of the most visited category per session 
# n_unique_category (int) - number of unique categories per session

if (project == 'dp' or project == 'he'):
    url_name = 'url1'
elif (project == 'pp' or project == 'ba'): 
    url_name = 'url2'

df_pageviews_category = df_pageviews[df_pageviews['pageview_category'] == True][['session_id', url_name]]

sessions_top_category = df_pageviews_category.groupby(['session_id'])[url_name].apply(lambda x: pd.Series.mode(x)[0]).to_frame()
sessions_top_category.rename(columns={url_name: 'top_category'}, inplace=True)

df_sessions = pd.merge(
    df_sessions,
    sessions_top_category,
    on='session_id',
    how='left'
)

sessions_n_top_category = df_pageviews_category.groupby(['session_id', url_name])[url_name].count().to_frame()
sessions_n_top_category.rename(columns={url_name: 'n_top_category'}, inplace=True)
sessions_n_top_category = sessions_n_top_category.sort_values(by=['n_top_category'], ascending=False)

df_sessions = pd.merge(
    df_sessions,
    sessions_n_top_category,
    on='session_id',
    how='left'
)

df_sessions.drop_duplicates('session_id', keep='first', inplace=True)
df_sessions = df_sessions.reset_index(drop=True)

sessions_n_unique_category = df_pageviews_category.groupby(['session_id'])[url_name].nunique().to_frame()
sessions_n_unique_category.rename(columns={url_name: 'n_unique_category'}, inplace=True)

df_sessions = pd.merge(
    df_sessions,
    sessions_n_unique_category,
    on='session_id',
    how='left'
)

df_sessions['referrer1'] = df_sessions['referrer'].str.extract(r'(\b(?<=:\/\/)\b\S+)') # referrer1 - everything behind ://
df_sessions['referrer2'] = df_sessions['referrer1'].str.extract(r'(\b(?<=\/)\b\S+[^\/])') # 2. part from referrer1 without / at the end
df_sessions['referrer3'] = df_sessions['referrer2'].str.extract(r'(\b(?<=\/)\b\S+[^\/])') # 2. part from referrer2 without / at the end

df_sessions['referrer1'] = df_sessions['referrer1'].str.extract(r'([^\/]*)') # 1. part from referrer1
df_sessions['referrer2'] = df_sessions['referrer2'].str.extract(r'([^\/]*)') # 1. part from referrer2

df_pageviews['referrer1'] = df_pageviews['referrer'].str.extract(r'(\b(?<=:\/\/)\b\S+)') # referrer1 - everything behind ://
df_pageviews['referrer2'] = df_pageviews['referrer1'].str.extract(r'(\b(?<=\/)\b\S+[^\/])') # 2. part from referrer1 without / at the end
df_pageviews['referrer3'] = df_pageviews['referrer2'].str.extract(r'(\b(?<=\/)\b\S+[^\/])') # 2. part from referrer2 without / at the end

df_pageviews['referrer1'] = df_pageviews['referrer1'].str.extract(r'([^\/]*)') # 1. part from referrer11
df_pageviews['referrer2'] = df_pageviews['referrer2'].str.extract(r'([^\/]*)') # 1. part from referrer2

# time when session starts
df_sessions['hour'] = pd.to_datetime(df_sessions['started_at'], unit='ms').dt.hour
df_sessions['day_name'] = pd.to_datetime(df_sessions['day']).dt.day_name()

df_pageviews_buy = df_pageviews[df_pageviews['pageview_buy'] == True]

df_pageviews_buy.loc[:,'buy_hour'] = pd.to_datetime(df_pageviews_buy['started_at'], unit='ms').dt.hour
df_pageviews_buy.loc[:,'buy_day'] = pd.to_datetime(df_pageviews_buy['started_at'], unit='ms').dt.day_name()

df_pageviews_buy = df_pageviews_buy.sort_values(by=['started_at'])

df_sessions = pd.merge(
    df_sessions,
    df_pageviews_buy[['session_id', 'buy_hour', 'buy_day']],
    on='session_id',
    how='left'
)

df_sessions.drop_duplicates('session_id', keep='first', inplace=True)

# 0 - other city,  1 - Bratislava,  2 - Košice, 3 - other regional city
city_BA = 'Bratislava'
city_KE = 'Košice'
city_regional = ['Trnava', 'Žilina', 'Nitra', 'Prešov', 'Banská Bystrica', 'Trenčín']

df_sessions.loc[df_sessions['city'] == city_BA, 'city_type'] = 1 
df_sessions.loc[df_sessions['city'] == city_KE, 'city_type'] = 2 
df_sessions.loc[df_sessions['city'].isin(city_regional), 'city_type'] = 3 

# pageview load speed
df_load.rename(columns={'at': 'load_time'}, inplace=True)

df_pageviews = pd.merge(
    df_pageviews,
    df_load[['pageview_id', 'load_time']],
    on='pageview_id',
    how='left'
)

df_pageviews.drop_duplicates('pageview_id', keep='first', inplace=True)
df_pageviews = df_pageviews.reset_index(drop=True)
df_load.rename(columns={'load_time': 'at'}, inplace=True)

# !. pageview load speed per session
df_pageviews = df_pageviews.sort_values(by=['started_at'])

df_sessions = pd.merge(
    df_sessions,
    df_pageviews[['session_id', 'load_time']],
    on='session_id',
    how='left'
)

df_sessions.drop_duplicates('session_id', keep='first', inplace=True)
df_sessions = df_sessions.reset_index(drop=True)

# session duration
session_duration = df_pageviews.groupby('session_id')[['duration']].sum()

df_sessions = pd.merge(
    df_sessions,
    session_duration,
    on='session_id',
    how='left'
)

numeric_values = [
    'n_product', 'n_non_product',
    'n_category', 'n_search', 'n_buy',
    'load_time',
    'n_top_product', 'n_top_category',
    'n_unique_product', 'n_unique_category',
    'city_type',
]

for column in numeric_values:
    df_sessions[column] = df_sessions[column].fillna(0)
    df_sessions[column] = df_sessions[column].astype(int)
    
string_values = [
    'continent', 'region', 'country', 'city',
    'browser.name', 'device.type', 'os.name',
    'referrer1', 'referrer2', 'referrer3',
    'top_product', 'top_category',
    'buy_hour', 'buy_day', 'day_name'
]

for column in string_values:
    df_sessions[column].fillna('-', inplace = True)

# repair session (Herbatica)

# pageviews_buy, where referrer1 = 'gate.gopay.cz', 'gopay.cz' & n_pageviews <= 2 - delete
df_pageviews_buy['referrer1'].unique() # 'gate.gopay.cz', 'gopay.cz' is redirection from payment

# df_pageviews_buy['referrer3'].unique() # krok-2 => correct redirection to purchase completion page

df_pageviews_buy['url3'].unique() # url3 should be NaN

# group by number of pageview_buy per session + list of purchase session with 1 pageview_buy
n_pv_buy = df_pageviews_buy.groupby('session_id')['pageview_buy'].sum().reset_index()
sessions_1_pv_buy = n_pv_buy[n_pv_buy['pageview_buy'] == 1]['session_id'].unique()

# list of pageview_buy where referrer = gopay and number of pageview_buy in session = 1 
pv_buy_1 = df_sessions[(df_sessions['session_id'].isin(sessions_1_pv_buy) & (df_sessions['referrer1'].isin(['gate.gopay.cz', 'gopay.cz'])))]['session_id'].unique()

sessions_fake_buy = df_sessions[(df_sessions['session_id'].isin(pv_buy_1)) & (df_sessions['n_pageviews'] < 3)]['session_id'].unique()
len(sessions_fake_buy)

# remove sessions with fake purchase
df_sessions = df_sessions[~df_sessions['session_id'].isin(sessions_fake_buy)]

# sessions, where referrer = gopay and more than 2 pageviews
sessions = df_sessions[df_sessions['session_id'].isin(pv_buy_1)]['session_id'].unique()

for s in sessions:
    df_pageviews = df_pageviews[~((df_pageviews['session_id'] == s) & (df_pageviews['referrer1'].isin(['gate.gopay.cz', 'gopay.cz'])))]
    df_sessions.loc[df_sessions['session_id'] == s, 'n_buy'] = 0
    df_sessions.loc[df_sessions['session_id'] == s, 'buy'] = False
    df_sessions.loc[df_sessions['session_id'] == s, 'buy_hour'] = '-'
    df_sessions.loc[df_sessions['session_id'] == s, 'buy_day'] = '-'    

### DF_USERS

# create df_users
df_users = df_sessions.groupby(['user_id','session_id'])[['n_buy']].sum()

df_users.loc[df_users['n_buy'] > 0, 'buy'] = 1
df_users.loc[df_users['n_buy'] == 0, 'buy'] = 0

# sum of values per sessions
temp = df_sessions.groupby(['user_id','session_id'])[
    ['n_pageviews', 'n_input', 'n_load', 'n_click', 'n_mouse_click', 'n_rage_click',
     'n_mouse_move','n_scroll_move','n_wild_mouse','n_scrandom',
     'n_events', 'n_product', 'n_non_product', 'n_category', 'n_filter', 'n_search', 'n_cart', 
     'effective_duration', 'duration']].sum()

df_users = pd.merge(
    df_users, 
    temp, 
    left_on=['user_id','session_id'],
    right_index=True, 
    how='left'
)

# number of sessions per user
temp = pd.DataFrame(df_sessions['user_id'].value_counts()) 
temp.rename(columns={'user_id': 'n_sessions'}, inplace=True) 

df_users = pd.merge(
    df_users, 
    temp, 
    left_on='user_id', 
    right_index=True, 
    how='left'
)

df_sessions = df_sessions.sort_values(by=['started_at'])

df_users = pd.merge(
    df_users,
    df_sessions[['user_id', 'session_id','country', 'city', 'continent', 'started_at', 'region', 'day', 'device.type', 'browser.name', 'os.name', 'referrer1', 'referrer2', 'referrer3', 'hour', 'day_name', 'city_type', 'load_time']],
    on=['user_id','session_id'],
    how='left'
)

df_users.drop_duplicates('session_id', keep='first', inplace=True)

len(df_users['user_id'].unique())

# time and day 1. pageview with completed purchase
df_pageviews_buy = df_pageviews[df_pageviews['pageview_buy'] == True]

df_pageviews_buy.loc[:,'buy_hour'] = pd.to_datetime(df_pageviews_buy['started_at'], unit='ms').dt.hour # transform timestamp to date time and select hour
df_pageviews_buy.loc[:,'buy_day'] = pd.to_datetime(df_pageviews_buy['started_at'], unit='ms').dt.day_name() # transform timestamp to date time and select day

df_pageviews_buy = df_pageviews_buy.sort_values(by=['started_at'])

df_users = pd.merge(
    df_users,
    df_pageviews_buy[['user_id', 'session_id','buy_hour', 'buy_day']],
    on=['user_id','session_id'],
    how='left'
)

df_users.drop_duplicates('session_id', keep='first', inplace=True)

# remove users where n_events = 0 and pageviews < 2
df_users = df_users[df_users['n_events'] != 0]
df_users = df_users[df_users['n_pageviews'] > 1]

# number of purchase sessions for each user
buy_count = df_users.set_index('user_id').buy.eq(1).sum(level=0).astype(int).reset_index()
buy_count = buy_count.rename(columns={'buy': 'buy_count'})

df_users = pd.merge(
    df_users,
    buy_count,
    on='user_id',
    how='left'
)

len(df_users['user_id'].unique())

df_users = df_users.sort_values('started_at')

df_users['mean_n_ses_between_buys'] = '-'
df_users['mean_time_between_buys'] = '-'

for user in df_users['user_id'].unique():
    find = False
    i = 0
    day_1 = 0
    day_2 = 0
    zeros = list()
    for session in df_users[df_users['user_id'] == user]['session_id']:
        buy = df_users[df_users['session_id'] == session]['buy'].values[0]
        if find == False and buy == 1:
            day_1 = datetime.strptime(df_users[df_users['session_id'] == session]['day'].values[0], "%Y-%m-%d") 
            find = True
        elif find == True and buy == 0:
            i += 1
        elif find == True and buy == 1:            
            day_2 = datetime.strptime(df_users[df_users['session_id'] == session]['day'].values[0], "%Y-%m-%d")
            zeros.append(i)
            i = 0
    df_users.loc[df_users['user_id'] == user, 'mean_n_ses_between_buys'] = sum(zeros)/len(zeros) if zeros else 0
    df_users.loc[df_users['user_id'] == user, 'mean_time_between_buys'] = abs((day_2 - day_1).days) if (day_2 != 0 and day_1 != 0) else 0

df_users['repeat_buyer'] = np.where(df_users['buy_count'] > 1, True, False)

len(df_users[df_users['repeat_buyer'] == 1]['user_id'].unique())

len(df_users['user_id'].unique())

# number of sessions between 1. and 2. purchase , if !2.purchase, select all sessions
df_users = df_users.sort_values('started_at')

df_users

# average number of days between 2 buys
print(df_users[df_users['day'] > "2020-06-02"]['mean_time_between_buys'].mean())
# median number of days between 2 buys
print(df_users[df_users['mean_time_between_buys'] > 0]['mean_time_between_buys'].median())
# min number of days between 2 buys
print(df_users[df_users['mean_time_between_buys'] > 1]['mean_time_between_buys'].min())
# max number of days between 2 buys
print(df_users[df_users['mean_time_between_buys'] < 362]['mean_time_between_buys'].max())

d1 = datetime.strptime(df_users[df_users['session_id'] == session]['day'].values[0], "%Y-%m-%d") 

df_users[(df_users['user_id'] == '1684b3ff-5f8d-4ebd-94cf-76f77a4abc0f') & (df_users['buy'] == 1)]

df_users[df_users['user_id'] == 'fc01d19f-a1a0-4bca-8ec0-a0bff986e7e3']

df_users[(df_users['mean_time_between_buys'] == 464) & (df_users['buy'] == 1)]

df_users_2 = pd.DataFrame()
for user in df_users['user_id'].unique():
    find = False
    for session in df_users[df_users['user_id'] == user]['session_id']:
        buy = df_users[df_users['session_id'] == session]['buy'].values[0]
        if find == False and buy == 1:
            find = True
            df_users_2 = df_users_2.append(df_users[df_users['session_id'] == session])
        elif find == True and buy == 0:
            df_users_2 = df_users_2.append(df_users[df_users['session_id'] == session])
        elif find == True and buy == 1: 
            df_users_2 = df_users_2.append(df_users[df_users['session_id'] == session])
            break

len(df_users_2['user_id'].unique())

# number of sessions between 1. and 2. purchase 
# number of sessions after 1. purchase 
n_sessions_after_1_buy = df_users_2.groupby('user_id')['session_id'].count().astype(int).reset_index()
n_sessions_after_1_buy.rename(columns={'session_id' : "n_sessions_after_1_buy"}, inplace = True)

df_users_2 = pd.merge(
    df_users_2,
    n_sessions_after_1_buy,
    on='user_id',
    how='left'
)

len(df_users_2['user_id'].unique())

df_users_2 = df_users_2[df_users_2['n_sessions_after_1_buy'] > 1]
df_users_2['mean_n_ses_between_buys'] = df_users_2['mean_n_ses_between_buys'].astype(float)

len(df_users_2['user_id'].unique())

# number of sessions after 1. purchase 
counter = collections.Counter(round(df_users_2[df_users_2['mean_n_ses_between_buys'] > 0 ]['mean_n_ses_between_buys']))
print(counter)

labels = counter.keys()
sizes = counter.values()

fig1, ax1 = plt.subplots()
patches, texts = ax1.pie(sizes, labels=labels, startangle=90)
plt.legend(labels, loc="best")
ax1.axis('equal')

plt.show()

# cut number of sessions, max avg session count between 2 purchases after 1. purchase
session_limit = round(df_users_2[df_users_2['mean_n_ses_between_buys'] > 0 ]['mean_n_ses_between_buys'].mean())
session_limit

temp = pd.DataFrame()

# selection_method = 1 # session with 1. purchase and sessions without purchase
selection_method = 2 # also session with 2. purchase

if selection_method == 1:
    # select sessions after 1. purchase, if session count < limit + 1, or limit + 2 - select all sessions
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] <= (session_limit + 1))].sort_values('started_at',ascending=True), ignore_index=False)
    df = df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] <= (session_limit + 2))].sort_values('started_at',ascending=True)
    temp = temp.append(df.drop(df.groupby('user_id').tail(1).index,axis=0), ignore_index=False)
    
    # select 1. purchase
    temp = temp.append(df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 2))].sort_values('started_at',ascending=True).groupby('user_id').head(1), ignore_index=False)
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 1))].sort_values('started_at',ascending=True).groupby('user_id').head(1), ignore_index=False)
        
    # select last x sessions (x = session limit)
    df = df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 2))].sort_values('started_at',ascending=False)
    df = df.drop(df.groupby('user_id').head(1).index,axis=0)
    temp = temp.append(df.groupby('user_id').head(session_limit), ignore_index=False)
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 1))].sort_values('started_at',ascending=False).groupby('user_id').head(session_limit), ignore_index=False)

if selection_method == 2:
    # select sessions after 1. purchase, if session count < limit + 1, or limit + 2 - select all sessions and 2.purchase
    temp = temp.append(df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] <= (session_limit + 2))].sort_values('started_at',ascending=True), ignore_index=False)
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] <= (session_limit + 1))].sort_values('started_at',ascending=True), ignore_index=False)

    # select 1. purchase
    temp = temp.append(df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 2))].sort_values('started_at',ascending=True).groupby('user_id').head(1), ignore_index=False)
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 1))].sort_values('started_at',ascending=True).groupby('user_id').head(1), ignore_index=False)

    # select last x sessions (x = session limit), or last sessions + 2.purchase
    temp = temp.append(df_users_2[(df_users_2['buy_count'] > 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 2))].sort_values('started_at',ascending=False).groupby('user_id').head(session_limit + 1), ignore_index=False)
    temp = temp.append(df_users_2[(df_users_2['buy_count'] <= 1) & (df_users_2['n_sessions_after_1_buy'] > (session_limit + 1))].sort_values('started_at',ascending=False).groupby('user_id').head(session_limit), ignore_index=False)


temp = temp.sort_values('started_at') 

df_users = temp
df_sessions = df_sessions[df_sessions['session_id'].isin(df_users_2['session_id'])]
df_pageviews = df_pageviews[df_pageviews['session_id'].isin(df_users_2['session_id'])]

len(df_users['user_id'].unique())

# number of visits of the most visited product and category per user and session
df_users = pd.merge(
    df_users,
    df_sessions[['session_id','top_product','n_top_product','n_unique_product','top_category','n_top_category','n_unique_category']],
    on='session_id',
    how='left'
)

# number of visits of the most visited product per user
if (project == 'dp' or project == 'he'):
    url_name = 'url2'
elif (project == 'pp' or project == 'ba'):
    url_name = 'url3'

df_pageviews_product = df_pageviews[df_pageviews['pageview_product'] == True][['user_id', url_name]] 

if (project == 'dp' or project == 'he' or project == 'ba'):
    users_top_product = df_pageviews_product.groupby(['user_id'])[url_name].apply(lambda x: pd.Series.mode(x)[0]).to_frame()
elif (project == 'pp'):
    users_top_product = df_pageviews_product.groupby(['user_id'])[url_name].agg(pd.Series.mode).to_frame()

users_top_product.rename(columns={url_name: 'top_product_u'}, inplace=True)

df_users = pd.merge(
    df_users,
    users_top_product,
    on='user_id',
    how='left'
)

users_n_top_product = df_pageviews_product.groupby(['user_id', url_name])[url_name].count().to_frame()
users_n_top_product.rename(columns={url_name: 'n_top_product_u'}, inplace=True)
users_n_top_product = users_n_top_product.sort_values(by=['n_top_product_u'], ascending=False)

df_users = pd.merge(
    df_users,
    users_n_top_product,
    on='user_id',
    how='left'
)

df_users.drop_duplicates('session_id', keep='first', inplace=True)
df_users = df_users.reset_index(drop=True)

users_n_unique_product = df_pageviews_product.groupby(['user_id'])[url_name].nunique().to_frame()
users_n_unique_product.rename(columns={url_name: 'n_unique_product_u'}, inplace=True)

df_users = pd.merge(
    df_users,
    users_n_unique_product,
    on='user_id',
    how='left'
)

# number of visits of the most visited category per user
if (project == 'dp' or project == 'he'):
    url_name = 'url1'
elif (project == 'pp' or project == 'ba'): # pp doesn't have categories
    url_name = 'url2'

df_pageviews_category = df_pageviews[df_pageviews['pageview_category'] == True][['user_id', url_name]]

users_top_category = df_pageviews_category.groupby(['user_id'])[url_name].apply(lambda x: pd.Series.mode(x)[0]).to_frame()
users_top_category.rename(columns={url_name: 'top_category_u'}, inplace=True)

df_users = pd.merge(
    df_users,
    users_top_category,
    on='user_id',
    how='left'
)

users_n_top_category = df_pageviews_category.groupby(['user_id', url_name])[url_name].count().to_frame()
users_n_top_category.rename(columns={url_name: 'n_top_category_u'}, inplace=True)
users_n_top_category = users_n_top_category.sort_values(by=['n_top_category_u'], ascending=False)

df_users = pd.merge(
    df_users,
    users_n_top_category,
    on='user_id',
    how='left'
)

df_users.drop_duplicates('session_id', keep='first', inplace=True)
df_users = df_users.reset_index(drop=True)

users_n_unique_category = df_pageviews_category.groupby(['user_id'])[url_name].nunique().to_frame()
users_n_unique_category.rename(columns={url_name: 'n_unique_category_u'}, inplace=True)

df_users = pd.merge(
    df_users,
    users_n_unique_category,
    on='user_id',
    how='left'
)

# number of add to cart clicks per session
if (project == 'he'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Pridať do košíka") | (df_click['args.text'] == "Do košíka") |
                                 (df_click['args.text'] == "Add to basket") | (df_click['args.text'] == "To basket") |
                                 (df_click['args.text'] == "Přidat do košíku") | (df_click['args.text'] == "Do košíku")]["session_id"].value_counts()) 
    temp.rename(columns={"session_id": "n_add_to_cart"}, inplace=True) 

if (project == 'pp'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Pridať do košíka") | (df_click['args.text'] == "Add to cart")]["session_id"].value_counts()) 
    temp.rename(columns={"session_id": "n_add_to_cart"}, inplace=True) 

if (project == 'ba'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Vložiť do košíka") | (df_click['args.text'] == "Do košíka")]["session_id"].value_counts()) 
    temp.rename(columns={"session_id": "n_add_to_cart"}, inplace=True)

df_sessions = pd.merge(df_sessions, temp, left_on='session_id', right_index=True, how='left')

df_users = pd.merge(
    df_users,
    df_sessions[['session_id','n_add_to_cart']],
    on='session_id',
    how='left'
)

df_users['n_add_to_cart'] = df_users['n_add_to_cart'].fillna(0)

# number of remove from cart clicks per session
if (project == 'he'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Odstrániť z košíka") | (df_click['args.text'] == "Odstranit z košíku")]["session_id"].value_counts()) 
    temp.rename(columns={"session_id": "n_remove_from_cart"}, inplace=True) 

if (project == 'pp'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Vymazať") | (df_click['args.text'] == "Delete")]["session_id"].value_counts())
    temp.rename(columns={"session_id": "n_remove_from_cart"}, inplace=True) 

if (project == 'ba'):
    temp = pd.DataFrame(df_click[df_click['args.text'] == "Odstrániť"]["session_id"].value_counts())
    temp.rename(columns={"session_id": "n_remove_from_cart"}, inplace=True) 

df_sessions = pd.merge(df_sessions, temp, left_on='session_id', right_index=True, how='left') 

df_users = pd.merge(
    df_users,
    df_sessions[['session_id','n_remove_from_cart']],
    on='session_id',
    how='left'
)

df_users['n_remove_from_cart'] = df_users['n_remove_from_cart'].fillna(0)

# number of next page clicks - only He, Ba - PP doesn't have next pages
pages = list(range(0, 500))
pages = [str(x) for x in pages]

if (project == 'he'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "Ďalšie produkty") | (df_click['args.text'] == "Ďalšia strana") |
                                (df_click['args.text'] == "Další produkty") | (df_click['args.text'] == "Další strana")]["session_id"].value_counts())
    temp.rename(columns={"session_id": "n_next_page"}, inplace=True) 


if (project == 'ba'):
    temp = pd.DataFrame(df_click[(df_click['args.text'] == "nasledujúca strana") | 
                                 (df_click['args.text'].isin(pages))]["session_id"].value_counts())
    temp.rename(columns={"session_id": "n_next_page"}, inplace=True) 

if (project == 'he') or (project == 'ba'):
    df_sessions = pd.merge(df_sessions, temp, left_on='session_id', right_index=True, how='left')

    df_users = pd.merge(
        df_users,
        df_sessions[['session_id','n_next_page']],
        on='session_id',
        how='left'
    )

    df_users['n_next_page'] = df_users['n_next_page'].fillna(0)

# transformation NaN to 0 and float to int
numeric_values = [
    'n_pageviews', 'n_input', 'n_load',
    'n_click', 'n_mouse_click',  'n_rage_click',
    'n_mouse_move', 'n_scroll_move', 'n_wild_mouse', 
    'n_scrandom',
    'n_events',
    'n_product', 'n_non_product', 'n_category',
    'n_filter', 'n_search',
    'n_cart', 'n_buy',
    'effective_duration', 'duration',
    'n_sessions',
    'hour',
    'city_type',
    'load_time',
    'n_top_product_u', 'n_unique_product_u',
    'n_top_category_u', 'n_unique_category_u',
    'n_add_to_cart', 'n_remove_from_cart', 
#     'n_next_page'
]

for column in numeric_values:
    df_users[column] = df_users[column].fillna(0)
    df_users[column] = df_users[column].astype(int)

# transformation NaN to '-' 
string_values = [
    'country', 'city', 'continent', 'region',
    'device.type', 'browser.name', 'os.name',
    'referrer1', 'referrer2', 'referrer3',
    'day_name', 
    'day',
    'buy_hour',
    'buy_day',
    'top_product_u', 'top_category_u',
]

for column in string_values:
    df_users[column].fillna('-', inplace = True)

len(df_users['user_id'].unique())

# export CSV 
# example filename: csv_he + df_users_metrics_ + 1 
df_users.to_csv(folder + '/df_users_metrics_' + str(selection_method) + '.csv', index=False, encoding='utf-8-sig')
df_sessions.to_csv(folder + '/df_sessions_metrics_' + str(selection_method) + '.csv', index=False, encoding='utf-8-sig')
df_pageviews.to_csv(folder + '/df_pageviews_metrics_' + str(selection_method) + '.csv', index=False, encoding='utf-8-sig')

# UNCHANGED
# df_inactivities
# df_input
# df_load
# df_click
# df_mouse_click
# df_rage_click
# df_mouse_move
# df_scroll_move
# df_wild_mouse
# df_scrandom
