==> scripts/7 - Transformation.ipynb <==
import numpy as np
import pandas as pd
import collections # counter

from sklearn import preprocessing
from sklearn.impute import SimpleImputer # missing data
from sklearn.preprocessing import StandardScaler # feature scaling
from sklearn.compose import ColumnTransformer

# folder = 'csv_he'
# folder = 'csv_pp_500'
folder = 'csv_b_500'
# selection_method = 1
selection_method = 2

df_users = pd.read_csv(folder + '/df_users_dom_' + str(selection_method) + '.csv')
df_pageviews = pd.read_csv(folder + '/df_pageviews_metrics_' + str(selection_method) + '.csv')
df_sessions = pd.read_csv(folder + '/df_sessions_metrics_' + str(selection_method) + '.csv')

pd.options.display.max_columns = None

columns = ['n_load', 'n_rage_click', 'n_wild_mouse', 'n_unique_product', 'n_unique_category',
           'mean_n_ses_between_buys', 'buy_count', 'n_buy', 'day', 'referrer2', 'referrer3', 'started_at', 'duration']
df_users = df_users.drop(columns, axis=1)

# n_next_page - only He
columns = ['n_pageviews', 'n_input', 'n_click', 'n_mouse_click', 'n_mouse_move',
           'n_scroll_move', 'n_scrandom', 'n_events', 'n_product', 'n_non_product', 
           'n_category', 'n_filter', 'n_search', 'n_cart', 'n_add_to_cart', 'n_remove_from_cart', 
           'n_next_page',
           'effective_duration', 'pv_product_mean_eff_duration', 'load_time']

data = pd.DataFrame()

# split each column to buy/no buy
for user in df_users['user_id'].unique():
    for column in columns:
        buy_true = 0
        buy_false = 0
        n_buy = 0
        n_no_buy = 0
        
        for session in df_users[df_users['user_id'] == user]['session_id']:
            val = df_users[df_users['session_id'] == session][column].values[0]
            buy = df_users[df_users['session_id'] == session]['buy'].values[0]
            if buy == 1:
                buy_true += val
                n_buy += 1
            if buy == 0:
                buy_false += val
                n_no_buy += 1
        df_users.loc[df_users['user_id'] == user, column + '_buy'] = round(buy_true/n_buy)
        df_users.loc[df_users['user_id'] == user, column + '_no_buy'] = round(buy_false/n_no_buy) if n_no_buy != 0 else 0
    df_users.loc[df_users['user_id'] == user,'num_of_sessions'] = n_no_buy + n_buy

len(df_users[(df_users['repeat_buyer'] == True) & (df_users['n_pageviews_no_buy'] == 0)]['user_id'].unique())

df_users = df_users.drop(columns,axis=1)

data = pd.DataFrame()
columns = ['country', 'city', 'city_type', 'continent', 'region', 'device.type', 'browser.name', 'os.name', 'referrer1']
for column in columns:    
    data[column] = df_users.groupby(['user_id'])[column].apply(lambda x: pd.Series.mode(x)[0])

data = data.reset_index()
df_users = df_users.drop(columns, axis = 1)

df_users = pd.merge(
    df_users,
    data,
    on='user_id',
    how='left'
)

df_users.columns

def set_buy_time(hour):
    if (hour > 2 and hour <= 10):
        return 1
    elif (hour <= 18):
        return 2
    else:
        return 3

def set_buy_day(day):
    weekday = ['Monday','Tuesday','Thursday','Wednesday','Friday']
    weekend = ['Saturday','Sunday']
    if (day in weekday):
        return 1
    else:
        return 2

df_users['buy_time'] = df_users['buy_hour'].apply(lambda x: 0 if x == '-' else set_buy_time(float(x)))
df_users['time'] = df_users['hour'].apply(lambda x: 0 if x == '-' else set_buy_time(float(x)))
df_users['buy_day'] = df_users['buy_day'].apply(lambda x: 0 if x == '-' else set_buy_day((x)) )
df_users['day'] = df_users['day_name'].apply(lambda x: 0 if x == '-' else set_buy_day((x)) )

columns = ['hour', 'day_name', 'buy_hour']
df_users = df_users.drop(columns,axis=1)

for i in df_users['time'].unique():
    df_users['buy_time_' + str(i)] = 0
    df_users['time_' + str(i)] = 0
    
for i in df_users['day'].unique():
    df_users['buy_day_' + str(i)] = 0
    df_users['day_' + str(i)] = 0

i = 0
columns = ['buy_time', 'time', 'buy_day', 'day']
for user in df_users['user_id'].unique():
    for column in columns:
        for session in df_users[df_users['user_id'] == user]['session_id']:
            val = df_users[df_users['session_id'] == session][column].values[0]
            if val != 0:
                df_users.loc[df_users['user_id'] == user, column + '_' + str(val)] += 1

df_users = df_users.drop(columns,axis=1)

columns = ['user_id', 'session_id', 'buy', 'top_product', 'n_top_product', 'top_category', 'n_top_category']
df_users = df_users.drop(columns,axis=1)

df_users = df_users.drop_duplicates(subset=None, keep='first', inplace=False, ignore_index=False)
len(df_users)

for column in df_users.columns:
    df_users[column].fillna('-',inplace=True)

df_users.to_csv(folder + '/df_users_transform_' + str(selection_method) + '.csv', index=False, encoding='utf-8-sig')
